requirements.txt
# Core dependencies for Stock Analysis Tool
yfinance
mcp>=0.1.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

smolagents
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0
httpx>=0.27.0
requests>=2.31.0
